requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
pydantic==2.8.2
python-dotenv==1.0.1
email-validator==2.2.0
imapclient==2.3.1
beautifulsoup4==4.12.3
html2text==2024.2.26
huggingface-hub==0.24.6
transformers==4.43.3
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
openai==1.43.0
streamlit==1.37.1
requests==2.32.3
python-multipart==0.0.9
python-dateutil==2.9.0.post0
